Задание 2: to_f; задание 3: добавлено условие для гипотинузы; задание 4: ошибка в формуле x1 (- заменен на +)

+++ lesson_1/02_area_of_triangle.rb
@@ -1,5 +1,5 @@
 print "Основание треугольника: "
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 print "Высота треугольника: "
-h = gets.chomp.to_i
+h = gets.chomp.to_f
 puts "Площадь треугольника: #{(a * h)/2}"
